antes del proyecto

+++ Back-End/src/main/java/com/porfolio/jeo/repositorio/IPersona.java
@@ -5,8 +5,10 @@ import com.porfolio.jeo.entidad.persona;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
 
-public interface IPersona extends JpaRepository<persona, Long>{
+public interface IPersona extends JpaRepository<persona, Integer>{
+    
     
 }
